[DEV-2.3] Fixed bugs regarding add garbage lost in singleplayer
