+++ model/TetrisModel.java
@@ -4,6 +4,7 @@ import javafx.animation.Animation;
 import javafx.animation.AnimationTimer;
 import javafx.animation.KeyFrame;
 import javafx.animation.Timeline;
+import javafx.application.Platform;
 import javafx.util.Duration;
 import multiplayer.Client;
 import views.ConnectView;
@@ -366,6 +367,9 @@ public class TetrisModel {
         }else { // If the game is a multiplayer game, tell the server that this player lost
             this.client.sendPacket(this.client.numConnections, true,true, 0);
             this.isMultiplayer = false;
+            Platform.runLater(() -> {
+                TetrisApp.view.gameView.createLostView();
+            });
         }
         gameOn = false;
     }
