+++ views/SettingsView.java
@@ -1,7 +1,5 @@
 package views;
 
-import javafx.event.Event;
-import javafx.event.EventType;
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
 import javafx.scene.Scene;
